Add control env vars, discard AUR deps in aurRPC

+++ PKGBUILD
@@ -3,49 +3,105 @@
 # Contributor: Randy Heydon <randy dot heydon at clockworklab dot net>
 # Contributor: saxonbeta <saxonbeta at gmail __com
 
-_fragment=${FRAGMENT:-#branch=devel}
+# matc is currently bugged: https://github.com/ElmerCSC/elmerfem/issues/306
+DISABLE_MATC=1
+# allow build with `intel-mkl`
 
-# Configuration.
-if (( DISABLE_TRILINOS ))
-  then _CMAKE_FLAGS+=(-DWITH_Trilinos=OFF)
-  else depends+=('trilinos'); _CMAKE_FLAGS+=(-DWITH_Trilinos=ON)
-fi
+# Configuration:
+# Use: makepkg VAR1=0 VAR2=1 to enable(1) disable(0) a feature
+# Use: {yay,paru} --mflags=VAR1=0,VAR2=1
+# Use: aurutils --margs=VAR1=0,VAR2=1
+# Use: VAR1=0 VAR2=1 pamac
+
+# Use FRAGMENT=#{commit,tag,brach}=xxx for bisect build
+_fragment=${FRAGMENT:-#branch=devel}
+# Use CMAKE_FLAGS=xxx:yyy:zzz to define extra CMake flags
+((CMAKE_FLAGS))      && mapfile -t -d: _CMAKE_FLAGS < <(echo -n "$CMAKE_FLAGS")
+
+((DISABLE_ALL)) && eval DISABLE_{TRILINOS,MMG,ELMERICE,CONTRIB,LUA,GUI,MPI}=1
+((DISABLE_TRILINOS)) && _use_trilinos=OFF || _use_trilinos=ON  # Disable Trilinos - linear system solver (Experimental)
+((DISABLE_MMG))      && _use_mmg=OFF      || _use_mmg=ON       # Disable MMG - dynamic remeshing
+((DISABLE_ELMERICE)) && _use_elmerice=OFF || _use_elmerice=ON  # Disable ElmerICE - glacier melting solver
+((DISABLE_CONTRIB))  && _use_contrib=OFF  || _use_contrib=ON   # Disable multishell solver for composite lamitanes
+((DISABLE_LUA))      && _use_lua=OFF      || _use_lua=ON       # Disable LUA scripting in solver definitions
+((DISABLE_MP))       && _use_openmp=OFF   || _use_openmp=ON    # Disable OpenMP threading
+
+((DISABLE_GUI))      && _use_elmergui=OFF || _use_elmergui=ON  # Disable ElmerGUI - QT GUI
+# Requires GUI
+((DISABLE_GUI))      && eval DISABLE_{GUILOG,GUITEST,MATC,OCC,PARAVIEW,QWT,VTK}=1
+((DISABLE_GUILOG))   && _use_guilog=OFF   || _use_guilog=ON    # Disable ElmerGUI Logger
+((DISABLE_GUITEST))  && _use_guitest=OFF  || _use_guitest=ON   # Disable ElmerGUI Tests
+((DISABLE_OCC))      && _use_occ=OFF      || _use_occ=ON       # Disable OCC - OpenCOLADA cad model import
+((DISABLE_MATC))     && _use_matc=OFF     || _use_matc=ON      # Disable MatC scripting in QT GUI
+((DISABLE_PARAVIEW)) && _use_paraview=OFF || _use_paraview=ON  # Disable ParaView - GUI post-process exporter
+((DISABLE_QWT))      && _use_qwt=OFF      || _use_qwt=ON       # Disable QWT - GUI convergence monitoring
+((DISABLE_VTK))      && _use_vtk=OFF      || _use_vtk=ON       # Disable VTK - GUI post-process Widget and exporter
+
+((DISABLE_MPI))      && _use_mpi=OFF      || _use_mpi=ON       # Disable OpenMPI parallelization
+# Require OpenMPI
+((DISABLE_MPI))      && eval DISABLE_{MUMPS,HYPRE}=1
+((DISABLE_MUMPS))    && _use_mumps=OFF    || _use_mumps=ON     # Disable Mumps - gausian elimination LAS solver
+((DISABLE_HYPRE))    && _use_hypre=OFF    || _use_hypre=ON     # Disable Hypre - multigrid LAS solver
+
 _CMAKE_FLAGS+=(
-        -DCMAKE_INSTALL_PREFIX=/usr \
-        -DCMAKE_BUILD_TYPE=Release \
-        -DELMER_INSTALL_LIB_DIR=/usr/lib \
-        -DWITH_CONTRIB=ON \
-        -DWITH_ELMERGUI=ON \
-        -DWITH_ELMERGUILOGGER=OFF \
-        -DWITH_ELMERGUITESTER=OFF \
-        -DWITH_ElmerIce=ON \
-        -DPHDF5HL_LIBRARY=/usr/lib/libhdf5_hl.so \
-        -DWITH_LUA=ON \
-        -DWITH_MATC=OFF \
-        -DWITH_MPI=ON \
-        -DWITH_OpenMP=ON \
-        -DWITH_QT5=ON \
-        -DWITH_Mumps=ON \
-        -DWITH_Hypre=ON \
-        -DWITH_OCC=ON \
-        -DWITH_VTK=ON \
-        -DNN_INCLUDE_DIR=/usr/include \
-        -DHYPRE_INCLUDE_DIR=/usr/include/hypre \
-        -DWITH_ScatteredDataInterpolator=ON \
-        -DWITH_PARAVIEW=ON
+        -DCMAKE_BUILD_TYPE=Release
+        -DCMAKE_INSTALL_PREFIX=/usr
+        -DELMER_INSTALL_LIB_DIR=/usr/lib
+
+        -DWITH_MPI=${_use_mpi}
+        -DWITH_Mumps=${_use_mumps}
+        -DWITH_Hypre=${_use_hypre}
+
+        -DWITH_OpenMP=${_use_openmp}
+
+        -DNN_INCLUDE_DIR=/usr/include
+        -DHYPRE_INCLUDE_DIR=/usr/include/hypre
+        -DPHDF5HL_LIBRARY=/usr/lib/libhdf5_hl.so
+        -DWITH_ScatteredDataInterpolator=ON
+
+        -DWITH_ELMERGUI=${_use_elmergui}
+        -DWITH_QT5=${_use_elmergui}
+        -DWITH_ELMERGUILOGGER=${_use_guilog}
+        -DWITH_ELMERGUITESTER=${_use_guitest}
+
+        -DWITH_MATC=${_use_matc}
+        -DWITH_LUA=${_use_lua}
+        -DWITH_OCC=${_use_occ}
+        -DWITH_QWT=${_use_qwt}
+        -DWITH_CONTRIB=${_use_contrib}
+        -DWITH_ElmerIce=${_use_elmerice}
+        -DWITH_VTK=${_use_vtk}
+        -DWITH_PARAVIEW=${_use_paraview}
+        -DWITH_Trilinos=${_use_trilinos}
 )
 
 pkgname=elmerfem-git
 _pkgname=elmerfem
-pkgver=9.0.r694.ge2946a83
+pkgver=9.0.r695.g6ab20a47
 pkgrel=1
 pkgdesc="A finite element software for multiphysical problems"
 arch=('x86_64')
 url="http://www.elmerfem.org"
 license=('GPL')
-depends+=('arpack' 'blas' 'qt5-script' 'python-pyqt5' 'qwt' 'netcdf-fortran-openmpi' 'paraview-opt' 'mumps-par' 'oce' 'vtk' 'hypre' 'mmg' 'libnn-git' 'libcsa-git' 'scalapack')
-depends+=('tbb' 'openmpi')
-depends+=('freetype2' 'qt5-base' 'fmt' 'glew' 'pugixml' 'libxcursor')
+
+# Core deps
+depends+=('arpack' 'blas' 'libnn-git' 'libcsa-git' 'scalapack')
+#depends+=('python-pyqt5')
+((!DISABLE_MPI))      && depends+=('netcdf-fortran-openmpi') || depends+=('netcdf-fortran')
+# Main repos
+((!DISABLE_GUI))      && depends+=('qt5-base' 'qt5-script' 'qt5-svg' 'glew')
+((!DISABLE_QWT))      && depends+=('qwt')
+((!DISABLE_VTK))      && depends+=('vtk' 'tbb' 'openmpi' 'freetype2' 'qt5-base' 'fmt' 'glew' 'pugixml' 'libxcursor')
+# AUR
+((!DISABLE_MMG))      && depends+=('mmg')
+((!DISABLE_TRILINOS)) && depends+=('trilinos')
+((!DISABLE_PARAVIEW)) && depends+=('paraview-opt') # paraview<>vtk conflict
+((!DISABLE_MUMPS))    && depends+=('mumps-par')    # mumps
+((!DISABLE_OCC))      && depends+=('oce')          # opencascade
+((!DISABLE_HYPRE))    && depends+=('hypre')
+((!DISABLE_MPI))      && depends+=('openmpi')
+((!DISABLE_MP))       && depends+=('openmp')
+
 makedepends=('git' 'gcc-fortran' 'cmake')
 provides=('elmerfem')
 conflicts=('elmerfem')
@@ -75,8 +131,9 @@ prepare() {
   cd "$srcdir/$_pkgname"
   sed -i 's/1 depth/1 ${depth}/g' fem/tests/CMakeLists.txt
   sed -i 's/FALSE/false/g' ElmerGUI/Application/vtkpost/matc.cpp
-  git revert --no-commit fb4f920711516254ae10c4edda576a2805787210
-  git apply -v "${srcdir}"/{arpack,FindMMG,print_target_properties,vtk9{.cmake,.1}}.patch
+  ((!DISABLE_VTK)) && patch+=(vtk9{.cmake,.1}.patch)
+  ((!DISABLE_MMG)) && patch+=({FindMMG,print_target_properties}.patch)
+  git apply -v "${srcdir}"/arpack.patch "${patch[@]/#/${srcdir}/}"
 }
 
 build() {
@@ -98,6 +155,7 @@ package() {
   make -C build DESTDIR="$pkgdir" install
   cd "$pkgdir/usr"
 
+if ((!DISABLE_GUI)); then
   # Remove unecessary libraries
   rm -rf -- lib/{*.a,ElmerGUI}
 
@@ -114,5 +172,6 @@ package() {
   cp share/ElmerGUI/edf-extra/* share/ElmerGUI/edf
   mv share/ElmerGUI/license_texts/GPL_EXCEPTION share/licenses/$_pkgname
   rm share/ElmerGUI/license_texts/*
+fi
 }
 # vim:set sw=2 ts=2 et:
